Update patch set 1
Patch Set 1:

> (2 comments)
 > 
 > > Unfortunately, the earlier version I showed you is not a viable
 > > solution.  It triggers a bug in all known versions of the Solaris
 > > linker that causes an invalid program header to be generated with
 > > some executables resulting in immediate segfault during linker
 > setup
 > > at execution.
 > 
 > Could you provide more details about the bug?

Symptom of bug is:

$ mdb ./a.exe
> ::run -v
mdb: stop on SIGSEGV
mdb: target stopped at:
ld.so.1`setup+0x6a5:    movzwq 0x36(%rax),%rcx
> $C
ffff80ffbffff3e0 ld.so.1`setup+0x6a5()
ffff80ffbffff4a0 ld.so.1`_setup+0x2e7()
ffff80ffbffff530 ld.so.1`_rt_boot+0x6c()
0000000000000002 0xffff80ffbffff810()

> What executables are problematic? Why?

At the moment, the TestCgoCrashHandler executable triggers the bug 100% of the time.  Because the bug itself is in ld(1), it's difficult to say what precise conditions trigger it.

As for why, the executable, was built by using ld's SEGMENT_ORDER directive to have it generate an executable using a desired layout, but as a result has bad program header information in some cases, which causes the runtime linker to access incorrect memory, and then die.

The bad program header information is due to a bug in how ld(1) calculates the base address, which is due to another bug caused by SEGMENT_ORDER which causes ld(1) to place the note segment in the middle of a PT_LOAD section where it doesn't belong.

Patch-set: 1
Label: Code-Review=0