Update patch set 1
Patch Set 1:

> Use Reloc.Variant. We can have two variants of R_ADRPOWER.
 > One for D-form and one for DS-form.

The rationale is I want to reduce the number of relocation stored
in memory, esp. for memory loads/stores, which are used a lot.

And logically, the pair of instructions should be treated as a single
unit for relocation purposes too. The ELF relocations are designed
to fit the limitation of linkers and to provide more flexibility than
needed, and we don't need to copy the over-complicated design.

PS: The relocation variant scheme is designed specifically to make
handling ELF relocations easier.

Patch-set: 1