Update patch set 1
Patch Set 1:

> > Use Reloc.Variant. We can have two variants of R_ADRPOWER.
 > > One for D-form and one for DS-form.
 > 
 > The rationale is I want to reduce the number of relocation stored
 > in memory, esp. for memory loads/stores, which are used a lot.
 
I hadn't thought about this angle.

I did some instrumentation: there are 64052 addrpower relocations when linking godoc. A relocation (in the linker) is 48 bytes, so doubling the number of relocations for a memory access would add about 3 megabytes to the heap of a process that uses about 210 megs, or about 1.5%. Not trivial, but not huge either.
 
 > And logically, the pair of instructions should be treated as a
 > single
 > unit for relocation purposes too.

I'm not sure I buy this. What if the go toolchain ever wanted to support more flexible instruction scheduling?

 > The ELF relocations are designed
 > to fit the limitation of linkers and to provide more flexibility
 > than
 > needed, and we don't need to copy the over-complicated design.
 > 
 > PS: The relocation variant scheme is designed specifically to make
 > handling ELF relocations easier.

You see the thing is, I know that this is what the relocation variant scheme is about but I find it completely baffling in practice. Maybe if it had some documentation, I'd be able to figure it out.

The model where one relocation modifies one instruction word, and a relocation type value always means the same thing has a certain simplicity to it, although it does lead to lots of repetitive code.

PS: we should really have this conversation on https://github.com/golang/go/issues/11374

Patch-set: 1