Update patch set 1
Patch Set 1:

> Abandoned
 > 
 > I'm going to look at fixing this in tip.

I'm not sure it is broken, the test is probably there for a reason.

It looks like the test is distinguishing ideal constant arithmetic and float64 arithmetic.  It's not clear to me why they would be different, but they are.  Maybe when converting from ideal to float64 it maps -0 to 0?  I'm not sure if that is correct or not.

We certainly get a literal -0 as the RHS of an assignment node.

Patch-set: 1