Update patch set 1
Patch Set 1:

(5 comments)

The code looks OK.  A few higher-level thoughts:

Why 6 numbers per node?  I'm not sure what all that is trying to distinguish.  To the extent that it is strict ancestor vs. possibly== ancestor, it is easy enough to distinguish those with an == test.  If that made the numbering more obvious, it would help understand this code.

I'd like to see a use of this before we check it in.  I get the feeling we could probably build a sparse tree directly instead of building the parent pointers in an array first and then building the sparse tree from that.

Patch-set: 1