Update patch set 3
Patch Set 3: Run-TryBot+1

Regarding the compilation time increase, perhaps we should decide dynamically, maybe a simple threshold, whether to use the old dom (which is probably faster for small n) or the new dom (which is definitely faster for large n). And perhaps, in a test mode controlled by a const flag, use both and compare.

Patch-set: 3