Update patch set 3
Patch Set 3:

Josh, the compilation time increase is for a different optimization that is intended to use the fancy numbering in clever ways.  It's heavier-weight than I like -- and also there's an infinite loop somewhere in tests, either running it, or in the generated code.

I think also that if we did eventually add it, it could run directly after CSE and take advantage of the just-done numbering, because CSE seems to not restructure the control flow graph.

Patch-set: 3