A+C: automated updates
The previous automated updates only added missing entries for commits
since Go 1.4, with the assumption that we were caught up on things
prior to that. (Initially motivated by the existence of junk email
address jokes in the early git history, which I initially didn't want
to whitelist)

But it turns out we were missing lots of stuff, at least for subrepos
prior to N months ago.

This is an update with all subrepos updated, with no date
restrictions.

Change-Id: I2b5580ae4f89ae9ba7eaa336cc54ce6d606c5379