Update patch set 1
Patch Set 1:

> VMOVUPS is an AVX SSE3 instruction ? (i'm not really sure). Is this
 > available on every amd64 CPU that we support ?

He's not actually using VMOVUPS, just MOVUPS.  The description is wrong (MOVUPS is SSE1).

But it brings up a question: why MOVUPS and not MOVOU?  The obvious reason is it uses one less byte to encode.  MOVOU currently translates to movdqu, maybe we should make it mean movups instead?  I don't see (or don't understand) what the downside of that would be.

Patch-set: 1