Update patch set 4
Patch Set 4:

This does not implement what was written in the issue. I asked for functions, not methods, and prefer that approach because most uses of Type and Value will not need them.

But that is something to discuss, and the discussion belongs in the issue tracker, not the code review. Please abandon this CL. The work itself is easy, it's the design and the decision that's hard.

Patch-set: 4
Reviewer: Gerrit User 5015 <5015@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0