Update patch set 2
Patch Set 2:

If v.ID were wild, then the valueMark check in checkFunc would panic.

In other builds, it dies in cse. In others, dies in peephole optimizer. In others, agenr.

I think that what is happening may be that we are miscompiling the compiler as a result of this change, and that results in a blow-up elsewhere.

I'm about out of time for today but will resume as soon as I can. CC'ing Todd just in case he wants to look. :)

Patch-set: 2