Update patch set 4
Patch Set 4:

Add this check back for now.  Then this CL will be good to submit.  You can then in a separate CL investigate whether removing it makes sense.

My understanding was that we often generate code which does a clearfat() of some structure, only to be immediately followed by writing real data to that same structure.  The compiler wants to elide the original zeroing, but it only has logic to do that if the zeroing was generated by componentgen.

The compiler can also elide initialization writes that are never read.  Again, only if componentgen generated the zero writes so the reads can be matched up to the writes.  An example:

type T struct {
	x, y, z int
}

func f(i int) int {
	var a = T{3, 4, 5}
	return a.y
}

Patch-set: 4
Label: Code-Review=0