encoding/xml: support for matching a fields namespace by identifier or url
To shed some light on the issue/fix, consider a RSS feed parser thats deals with
namespaces from a variety of definitions. I could expect a few different kinds of
xmlns definitions for the same type of structure. ie. consider mRSS feeds in the
wild that use the "media" namespace, you will find:

Xmlns wasn't defined, but the namespace was used (ie. for mRSS with media namespace)
Xmlns was defined as xmlns:media="http://search.yahoo.com/mrss/"
Xmlns was defined as xmlns:media="http://search.yahoo.com/mrss"

I noticed that encoding/xml would track the xmlns' in a map to the url, and would
match the struct tags to the url. The issue of course here is with 2 and 3,
where the difference between a "/" would throw off the parser.

This is a simple fix that will match a namespace for the field struct tag
defintion by the namespace identifier or url.

Fixes #12624.

Change-Id: I6f95e20bb5172fe219ad9f0cc83d369b3e831a6e