bufio: allow Scanner to accept a user-provided buffer
By analogy with io.CopyBuffer, add Scanner.Buffer, which lets
the user give a buffer to the scanner. This buffer will not be
reallocated; it is the caller's responsibility to provide a big enough
buffer for whatever is required. In return, Scan guarantees it
will not allocate more storage.

We call it Buffer not SetBuffer for consistency with Split, which
perhaps should have been called SetSplit; too late regardless.

Both Buffer and Split panic if they are called after Scan. The
panic in Split is new, but the comment on the method already
said it needed to be called first, so we might as well add the
verification while we're doing it for Buffer.

This method allows precise user control of storage at the cost of more
predetermination.

Fixes #11702.

Change-Id: I80e3d0e3830562fdabd4f7b08f322e1378248c39

Change-Id: I955ff939426eecf499941e5324f330dfb4632ff5