archive/tar: convert Next() to be loop based
Though not specified by any format, most of the tar utilies like GNU,
BSD, and STAR store the latest meta header rather than the first.
Iterating through the headers using recursion makes this harder to
do and also makes it harder to follow the logic.

Furthermore, extract the logic for creating regFileReader and
sparseFileReader objects into their own handler functions that
are only called from Next. The goal is to eventually get to world
where the only methods that change the Reader state are Read, Next,
skipUnread, handleRegularFile, and handleSparseFile.

All of the other methods currently attached to Reader only do so
because the way they report errors is by setting tr.err (which
often goes unchecked on accident). We want to avoid this in the future
since it makes error analysis really difficult. That is a future CL.

Also, it may seem pointless to have seperate handleRegularFile and
handleSparseFile functions, but they will have more logic in them.
I did not place the error checking logic that is needed for various
other issues. That is a future CL.

Fixes #12638

Change-Id: Ic1557256fc1363c5cb26570e5d0b9f65a9e57341