Update patch set 2
Patch Set 2:

This is a significant change to the output that actually deletes information that used to be present. That is not an improvement. We already have runtime.Stack; providing just another way to get the same output does not add functionality.

Please roll this back and do it over, providing a version that maintains the format.

Patch-set: 2
Reviewer: Gerrit User 5015 <5015@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0