Update patch set 2
Patch Set 2:

> This is a significant change to the output that actually deletes
 > information that used to be present. That is not an improvement. We
 > already have runtime.Stack; providing just another way to get the
 > same output does not add functionality.
 > 
 > Please roll this back and do it over, providing a version that
 > maintains the format.

The function was deprecated, now it isn't. That is why I saw this CL as an improvement. Furthermore the original only offered to attempt program source recovery, so not returning anything is within the realm of possible outputs.

Anyhow, it's not clear to me how to usefully preserve the format without making the implementation more complex than the original. I'll roll this back if you want of course, but I don't see another reasonable way forward.

Patch-set: 2