archive/tar: forbid NUL character in USTAR and GNU fields
USTAR and GNU strings are NUL-terminated. Thus, we should never
allow the NUL terminator, otherwise we will lose data round-trip.

Relevant specification text:
<<<
The fields magic, uname, and gname are character strings each terminated by a NUL character.
>>>

Furthermore, we reject NULs in PAX keys, and NULs for PAX values
representing the USTAR string fields.

Technically, PAX keys and values should be UTF-8, but the observance
of invalid files in the wild causes us to be more liberal.
<<<
The <length> field, <blank>, <equals-sign>, and <newline> shown shall
be limited to the portable character set, as encoded in UTF-8.
>>>

Change-Id: I305b794d9d966faad852ff660bd0b3b0964e52bf