spec: clarify floating-point conversions of numeric constants
The spec defines precise numeric constants which do not overflow.
Consequently, +/-Inf and NaN values were excluded. The case was not
explicit for -0.

Clarified by explicitly mentioning that there are no -0 constants.
Also removed references to package math (which does not belong into
the spec and may be misleading in the respectice section).

That said, the conversion rules explicitly say that T(x) (for a numeric
x and floating-point type T) is the value obtained from rounding x per
IEEE 754. Such rounding can produce -0 which is a value of T (float32
or float64), but we don't support (typed or untyped) -0 constants.

Clarify by adding a sentence to the conversion rules.

For more details, see the issue below.

Fixes #12576.

Change-Id: Ibe3c676372ab16d9229f1f9daaf316f761e074ee