archive/tar: make Writer error handling consistent
Motivation for change:
* Writer.Write does not check if tw.err was already set
* Writer.Fush does not check if either tw.err or tw.closed was set

Thus, to make the error handling more consistent we want to check
those conditions before performing the core work of any of those
methods. As a simplification, we can eliminate tw.closed and fold
that logic into tw.err, by setting tw.err to ErrWriteAfterClose.
This ensures that calls to WriteHeader, Write, and Flush all
return that error after Close is called.

Change-Id: If433b3c95dcce1801c9c6e3b93749986797a4611