compress/flate: remove use of ReadError
A vast majority of the time, ReadError isn't even returned during
IO operations. Instead, an unwrapped error will be returned because
of the ReadByte call on L701.

Although this is technically an API change, any user reliant on
this error would not have worked properly anyways due to the fact
that most IO error are not wrapped. We might as well remove use
of ReadError. It is useless and actually makes clients that do
depend on catching IO errors more difficult.

Fixes #11856

Change-Id: Ib5fec5ae215e977c4e85de5701ce6a473d400af8