cmd/doc: don't stop after first package if the symbol is not found
The test case is
	go doc rand.Float64
The first package it finds is crypto/rand, which does not have a Float64.
Before this change, cmd/doc would stop there. After this change, we get:

% go doc rand.Float64
package rand // import "math/rand"

func Float64() float64

    Float64 returns, as a float64, a pseudo-random number in [0.0,1.0) from the
    default Source.
%

Another nice consequence is that if a symbol is not found, we might get
a longer list of packages that were examined:

% go doc rand.Int64
doc: no symbol Int64 in packages crypto/rand, math/rand
exit status 1
%

This change introduces a coroutine to scan the file system so that if
the symbol is not found, the coroutine can deliver another path to try.
(This is darned close to the original motivation for coroutines.)

The channel on which the messages are delivered is unbuffered, so
the goroutine is throttled. We set GOMAXPROCS=1 so that
once a path is discovered, the main goroutine will process it before
the scanner starts up again. Concurrency is not parallelism.

Turns out it also cleans up a bit of a mess in stopping processing early.

Change-Id: I2789505b9992cf04c19376c51ae09af3bc305f7f