Update patch set 3
Patch Set 3:

> Uploaded patch set 3.

I hope this is better, I've put the constants beside the usage, as that seemed to be what was done elsewhere. Although statusLine() seemed almost suitable for generating error messages, I've avoided using it  because it would require modifying that code and we don't have a request to send it, if we did adapt it to not require a request and not add the proto we could use it to get the error messages instead.

I wasn't clear on whether you would like to ditch the function sendError, if you prefer I can easily get rid of this new function and just put io.WriteString inline at each error.

Patch-set: 3