Update patch set 3
Patch Set 3: Code-Review+1

Just curious if there are specific changes I should make here and if we're happy with this approach, or if we wanted to go with one of the other approaches discussed in #9763. Let me know if I should indeed make changes to this.

Patch-set: 3
Reviewer: Gerrit User 5683 <5683@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+1