runtime: update gcController.scanWork regularly
Currently, gcController.scanWork is updated irregularly since it is
only read at the end of the GC cycle. We're about to read it during
the GC cycle to improve the assist ratio revisions, so modify gcDrain*
to regularly flush to gcController.scanWork in much the same way we
currently regularly flush to gcController.bgScanCredit.

One consequence of this is that it's difficult to keep gcw.scanWork
monotonic, so we give up on that and simply return the amount of scan
work done by gcDrainN rather than calculating it in the caller.

Change-Id: I7b50acdc39602f843eed0b5c6d2dacd7e762b81d