runtime: directly track GC assist balance
Currently we track the per-G GC assist balance as two monotonically
increasing values: the bytes allocated by the G this cycle (gcalloc)
and the scan work performed by the G this cycle (gcscanwork). The
assist balance is hence assistRatio*gcalloc - gcscanwork.

This works, but has two important downsides:

1) It requires floating-point math to figure out if a G is in debt or
   not. This makes it inappropriate to check for assist debt in the
   hot path of mallocgc, so we only do this when a G allocates a new
   span. As a result, Gs can operate "in the red", leading to
   under-assist and extended GC cycle length.

2) Revising the assist ratio during a GC cycle can lead to an "assist
   burst". If you think of a plot of scan work performed versus heap
   allocated, the assist ratio is the slope of the desired line.
   However, in the current system, this "line" always passes through
   the origin, so if the runtime increases the assist ratio, the line
   will "jump" up to this new line at the current heap size. Hence,
   the next time each G assists, it has to catch up to this new line.

This commit replaces this approach with directly tracking the GC
assist balance in terms of allocation bytes. Allocating N bytes simply
decreases this by N and assisting raises it by the amount of scan work
performed divided by the assist ratio (to get back to bytes).

This will make it cheap to figure out if a G is in debt, which will
let us efficiently check if an assist is necessary *before* performing
an allocation and hence keep Gs "in the black".

This also fixes assist bursts because the assist ratio is now in terms
of *remaining* work, rather than work from the beginning of the GC
cycle. Hence, the assist "line" becomes continuous, with the slope
continuously revised as the heap size increases to keep the assists on
target.

Change-Id: Ia821c5f07f8a433e8da7f195b52adfedd58bdf2c