cmd/internal/obj, runtime: add NOFRAME flag to suppress stack frame set up on ppc64x
DO NOT SUBMIT (needs docs, do we want to change arm64 and arm in the same change?)

Replace the confusing game where a frame size of $-8 would suppress the
implicit setting up of a stack frame with a nice explicit flag.

The code to set up the function prologue is still a little confusing but better
than it was.

Change-Id: I1d49278ff42c6bc734ebfb079998b32bc53f8d9a