Update patch set 2
Patch Set 2:

This change surprises me because I can't understand how
could reordering the fields in ProgInfo affect the size of Prog
or ProgInfo.

But I did an experiment, and found out reorder the fields in
ProgInfo does reduce the size of ProgInfo by 8. This doesn't
make sense.

// the original ProgInfo
type ProgInfo struct {
    Flags    uint32 // flag bits
    Reguse   uint64 // registers implicitly used by this instruction
    Regset   uint64 // registers implicitly set by this instruction
    Regindex uint64 // registers used by addressing mode
    _        struct{} // to prevent unkeyed literals
}

If we move the Flags field to after the Regindex field, then the
size reduces from 40 bytes to 32 bytes. Why?

Further experiment showed that removing the empty struct{}
at the end also achieve the same result.

So the conclusion is that: contrary to popular belief, adding
an empty field of struct{} can indeed make the struct larger.

I think this is a bug in the compiler. For example, see
http://play.golang.org/p/OTr-GLz2UN

Actually, this is a regression from Go 1.4. Possibly due to the
c2go conversion. I will take a closer look.

Patch-set: 2
Reviewer: Gerrit User 5055 <5055@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0