Update patch set 2
Patch Set 2:

I'm not sure I quite understand your comments. It's quite intentional that both of these are in bytes rather than scan work units. Tracking these in bytes avoids having to use floating point math in the hot path of mallocgc. It doesn't give up any of the flexibility of scan work units because when we go to do an assist, we convert bytes in to scan work using the assist ratio and then do the assist based on scan work. We could track the assist debt in units of scan work, but the math is virtually identical; we're just moving around where we convert between allocation bytes and scan work.

Patch-set: 2