DO NOT REVIEW: text/template: allow reading fields of non-empty interfaces
Read what a non-empty interface points to.

The deleted lines were added in https://codereview.appspot.com/4810060/,
which attempted to break an infinite loop. That was a long time ago.
If I just delete these lines with current codebase, the test "bug1"
(added in that CL) does not fail.

I've added a new check in case v.Elem()==v.

Change-Id: I224eac95b410124bd7894c099cfa2590065d99a2