net/http: don't hang if RemoteAddr() blocks
The PROXY protocol is supported by several proxy servers such as haproxy
and Amazon ELB.  This protocol allows services running behind a proxy to
learn the remote address of the actual client connecting to the proxy,
by including a single textual line at the beginning of the TCP
connection.
http://www.haproxy.org/download/1.5/doc/proxy-protocol.txt

There are several Go libraries for this protocol (such as
https://github.com/armon/go-proxyproto), which operate by wrapping a
net.Conn with an implementation whose RemoteAddr method reads the
protocol line before returning. This means that RemoteAddr is a blocking
call.

Before this change, http.Serve called RemoteAddr from the main Accepting
goroutine, not from the per-connection goroutine. This meant that it
would not Accept another connection until RemoteAddr returned, which is
not appropriate if RemoteAddr needs to do a blocking read from the
socket first.

This commit moves the RemoteAddr call (in srv.newConn) into the
per-connection goroutine.  It also simplifies srv.newConn to not return
an error (because it never does).

It makes one tiny semantic change: if the constant
debugServerConnections is changed to true and the ConnState hook is
used, the StateNew invocation of the ConnState will be passed the
underlying connection (c.w) rather than the wrapped logging
connection (c.rwc).

Change-Id: Ie53933c3a57531c5ed71e3abbf221ff010324e86