Update patch set 3
Patch Set 3:

The problem is I cannot figure out when to do them -- or are you saying I just add a (use of a) field to OADDR, and tweak the race detector code to set that, and then be conditional on that?  I was about to fire up the existing compiler under a debugger to figure out WTF it is doing with these.

Patch-set: 3