runtime: cache two workbufs to reduce contention
DO NOT SUBMIT. The worst case scenario this defends against is real,
but also quite difficult to trigger. We should decide whether or not
we want to do this. The cost is some code complexity, though not a
lot. TODO: Performance benchmarks.

Currently the gcWork abstraction caches a single work buffer. As a
result, if a worker is putting and getting pointers right at the
boundary of a work buffer, it can flap between work buffers and
(potentially significantly) increase contention on the global work
buffer lists.

This change modifies gcWork to instead cache two work buffers and
switch off between them. This introduces one buffers' worth of
hysteresis and eliminates the above performance worst case by
amortizing the cost of getting or putting a work buffer over at least
one buffers' worth of work.

In practice, it's difficult to trigger this worst case with reasonably
large work buffers. On the garbage benchmark, this reduces the max
writes/sec to the global work list from 32K to 25K and the median from
6K to 5K. However, if a workload were to trigger this worst case
behavior, it could significantly drive up this contention.

Change-Id: I68bcbf74ccb716cd5e844a554f67b679135105e6